fix: translations of languages with a region subtag
