+++ .changeset/brave-kings-vanish.md
@@ -1,0 +1,5 @@
+---
+'@astrojs/starlight': patch
+---
+
+Fixes a UI string translation issue for languages with a region subtag.
