+++ packages/starlight/__tests__/test-utils.ts
@@ -52,8 +52,9 @@ function mockDoc(
 
 function mockDict(id: string, data: z.input<ReturnType<typeof i18nSchema>>) {
 	return {
-		id,
+		id: project.legacyCollections ? id : id.toLocaleLowerCase(),
 		data: i18nSchema().parse(data),
+		filePath: project.legacyCollections ? undefined : `src/content/i18n/${id}.yml`,
 	};
 }
