+++ packages/starlight/utils/navigation.ts
@@ -13,7 +13,12 @@ import type {
 import { createPathFormatter } from './createPathFormatter';
 import { formatPath } from './format-path';
 import { BuiltInDefaultLocale, pickLang } from './i18n';
-import { ensureLeadingSlash, ensureTrailingSlash, stripLeadingAndTrailingSlashes } from './path';
+import {
+	ensureLeadingSlash,
+	ensureTrailingSlash,
+	stripExtension,
+	stripLeadingAndTrailingSlashes,
+} from './path';
 import { getLocaleRoutes, routes, type Route } from './routing';
 import { localeToLang, localizedId, slugToPathname } from './slugs';
 import type { StarlightConfig } from './user-config';
@@ -508,12 +513,6 @@ function applyPrevNextLinkConfig(
 	}
 	// Otherwise, if the global config is enabled, return the generated link if any.
 	return paginationEnabled ? link : undefined;
-}
-
-/** Remove the extension from a path. */
-function stripExtension(path: string) {
-	const periodIndex = path.lastIndexOf('.');
-	return path.slice(0, periodIndex > -1 ? periodIndex : undefined);
 }
 
 /** Get a sidebar badge for a given item. */
