+++ packages/starlight/virtual-internal.d.ts
@@ -1,16 +1,3 @@
-declare module 'virtual:starlight/project-context' {
-	const ProjectContext: {
-		root: string;
-		srcDir: string;
-		trailingSlash: import('astro').AstroConfig['trailingSlash'];
-		build: {
-			format: import('astro').AstroConfig['build']['format'];
-		};
-		legacyCollections: boolean;
-	};
-	export default ProjectContext;
-}
-
 declare module 'virtual:starlight/git-info' {
 	export function getNewestCommitDate(file: string): Date;
 }
